Cleanup the course progress classes related to tables
